refactor: 구간 추가/삭제 로직 Sections로 이동
